Find candidates skipped exons

computes similarity between positive and negative GenomicRanges and then finds candidate exons in neg whose adjacent exons exist in the positive set and are also adjacent

Usage

find_candidates(gr_pos, gr_neg, id_pos, id_neg)

Examples

if (FALSE) { # \dontrun{
gr <- se_mock_data()
library(ggplot2)
library(GenomicRanges)
library(plyranges)
plot_gr(gr)
gr_pos <- filter(gr, coefs > 0)
gr_neg <- filter(gr, coefs < 0)
find_candidates(gr_pos, gr_neg, gr_pos$tx_id, gr_neg$tx_id)
} # }